kernel 3.10.0-1160.42.2.rt56.1182.el7.x86_64(for centos, from rpmbuild)

+++ drivers/acpi/acpica/achware.h
@@ -5,7 +5,7 @@
  *****************************************************************************/
 
 /*
- * Copyright (C) 2000 - 2017, Intel Corp.
+ * Copyright (C) 2000 - 2013, Intel Corp.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -66,9 +66,9 @@ acpi_status
 acpi_hw_validate_register(struct acpi_generic_address *reg,
 			  u8 max_bit_width, u64 *address);
 
-acpi_status acpi_hw_read(u64 *value, struct acpi_generic_address *reg);
+acpi_status acpi_hw_read(u32 *value, struct acpi_generic_address *reg);
 
-acpi_status acpi_hw_write(u64 value, struct acpi_generic_address *reg);
+acpi_status acpi_hw_write(u32 value, struct acpi_generic_address *reg);
 
 struct acpi_bit_register_info *acpi_hw_get_bit_register_info(u32 register_id);
 
@@ -127,7 +127,7 @@ acpi_hw_clear_gpe_block(struct acpi_gpe_xrupt_info *gpe_xrupt_info,
 
 acpi_status
 acpi_hw_get_gpe_status(struct acpi_gpe_event_info *gpe_event_info,
-		       acpi_event_status *event_status);
+		       acpi_event_status * event_status);
 
 acpi_status acpi_hw_disable_all_gpes(void);
